// src/models/types/NodeType.ts
export enum NodeType {
    IMAGE = 'image',
    COMPANY = 'company',
    BRAND = 'brand',
    CAR_MODEL = 'car model',
    CAR_MODEL_VARIANT = "car model variant",
    RACE_TRACK = 'race track',
    TRACK_LAYOUT = 'track layout',
    RACING_SERIES = "racing series",
    RACING_EVENT = "racing event",
    RACING_SESSION = "racing session",
    SESSION_RESULT = "session result",
    LAP_TIME = "lap time",
    RACING_GAME = "racing game",
    GAMING_PLATFORM = "gaming platform",
    NODE = 'node',
}